requirements.txt
mistralai>=1.6.0
prompt-toolkit>=3.0.36
rich>=13.3.2
pyyaml>=6.0
pygments>=2.15.0
pytest>=7.3.1
click>=8.1.3
python-dotenv>=1.0.0
